ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
openai==1.59.4
anthropic==0.40.0
elevenlabs==1.12.0
pytrends==4.9.2
openai-whisper==20240930
moviepy==2.1.1
ffmpeg-python==0.2.0
Pillow==11.1.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
boto3==1.35.91
supabase==2.11.0
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.11
aiofiles==24.1.0
assemblyai==0.35.0
